Correct NativeAOT info
Context: 799057175121b968d11872e034bd8226e8a076e1

Commit 79905717 was likely "too cute" with
`Scratch.UnoSkiaAndroidNativeAOT.csproj`, as `$(PublishAOT)` was not
set under the .NET 10 build.

This could be verified by viewing the `unzip -l` output.

.NET 9 and .NET 10 *without* `$(PublishAOT)`=true:

	% ls -lh bin/Release/net9.0-android/android-x64/Scratch.UnoSkiaAndroidNativeAOT-Signed.apk
	-rw-rw-rw-  1 jon  staff    30M Jan  1  1981 bin/Release/net9.0-android/android-x64/Scratch.UnoSkiaAndroidNativeAOT-Signed.apk

	% unzip -l bin/Release/net9.0-android/android-x64/Scratch.UnoSkiaAndroidNativeAOT-Signed.apk
	…
	     7744  01-01-1981 01:01   lib/x86_64/libaot-CommonServiceLocator.dll.so
	   502760  01-01-1981 01:01   lib/x86_64/libmonodroid.so
	  3270136  01-01-1981 01:01   lib/x86_64/libmonosgen-2.0.so
	   166264  01-01-1981 01:01   lib/x86_64/libxamarin-app.so
	…

The presence of `libaot-*`, `libmonodroid.so`, `libmonosgen-2.0.so`
and `libxamarin-app.so` indicates that NativeAOT is *not* used.

Compare to .NET 10 *with* `$(PublishAOT)`=true:

	% ls -lh bin/Release/net10.0-android/android-x64/*-Signed.apk
	-rw-rw-rw-  1 jon  staff    40M Jan  1  1981 bin/Release/net10.0-android/android-x64/Scratch.UnoSkiaAndroidNativeAOT-Signed.apk

It's ~10MB larger.

	% unzip -l bin/Release/net10.0-android/android-x64/Scratch.UnoSkiaAndroidNativeAOT-Signed.apk | grep .so$
	  2734832  01-01-1981 01:01   lib/x86_64/libHarfBuzzSharp.so
	 64483040  01-01-1981 01:01   lib/x86_64/libScratch.UnoSkiaAndroidNativeAOT.so
	  9602656  01-01-1981 01:01   lib/x86_64/libSkiaSharp.so
	  6450936  01-01-1981 01:01   lib/x86_64/libc++_shared.so

Those are *all* of the native libraries.  ***No*** `libmonosgen-2.0.so`.

Additionally, Displayed time is *much* shorter: less than 1 second,
vs. nearly 5s with MonoVM.  (On an x64 emulator.  YMMV.)

Update `README.md` so that `-p:PublishAOT=true` is explicitly provided
on the command-line, ensuring that `$(PublishAOT)`=true.